Revert "Removed setInterval and replaced with new timer module"
