+++ index.html
@@ -5,11 +5,11 @@
     <link rel="stylesheet" type="text/css" href="styles/main.css" />
 </head>
 <body>
-    <div id="face"></div>
+    <div id="eye"></div>
     <script>
         if (!Array.from) Array.from = require('array-from');
         require('behaviorify/register');
-        require('./src');
+        require('./src/main');
     </script>
 </body>
 </html>
